requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
diffusers==0.10.0
transformers==4.25.1
fastapi==0.88.0
torchvision==0.13.1
ftfy==4.4.3
accelerate==0.15.0
uvicorn==0.20.0